refactor: optimize Chest component structure and loading behavior for improved performance
